Fetch mccode-pooch tags
The workflow to update the tagged registries relies on comparing tags present in this and the main McCode repositories. It could not do so successfully before since it did not *checkout* the tags in both repositories.

In the previous state, all tags found in the main repository were marked as missing in this repository, which was incorrect.
The registry files were then regenerated for every tag, and a failure was encountered when attempting to push the 'new' tags to the remote repository.